requirements.txt
Django==5.2.10
djangorestframework==3.16.1
django-filter==25.2
Markdown==3.10
django-environ==0.11.2
mysqlclient==2.2.6
drf-spectacular==0.29.0